rename vault to mapvault
