+++ res/mapvault/map_icons/army.png
