+++ res/mapvault/map_icons/hydro.png
