+++ res/mapvault/map_icons/mass.png
